blog/modules/user/serializers/user.py
from rest_framework import serializers
from blog.modules.user.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'name']

