requirements.txt
asgiref==3.3.1
attrs==20.3.0
aws-xray-sdk==2.6.0
backcall==0.2.0
bleach==3.2.1
botocore==1.19.49
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
future==0.18.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.10.0
json5==0.9.4
jsonpickle==1.4.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
jupyterlab==2.1.2
jupyterlab-server==1.1.3
MarkupSafe==1.1.1
mistune==0.8.4
mysqlclient==2.0.3
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.8
pandas==1.0.3
pandocfilters==1.4.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.15.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
tornado==6.0.3
tqdm==4.46.0
traitlets==5.0.5
urllib3==1.25.9
visidata==1.5.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
